Plotting an interactive diversity map in a hierarchy

Plot an interactive map of hierarchy of diversity, from local population to regions and ecosystem with self-defined hierarchy structure. This function plots the diversity in the hierarchy, from local population to regions and ecosystem with self-defined hierarchical structure.

PlothieD(x, size = "Dq2", color_min = "hsl(56,80%,80%)", 
color_max = "hsl(341,30%,40%)", ...)

Arguments

x
Genetic diversity files, HierDgenind object, produced from function "HierDgenind".

size
The size of the diversity or the value of diversity will be plotted. This can be choosen after calculated the diversity profiles using HierDgenind.

color_min
string representing the minimum value of the color range for the circles. The string can be either a hexadecimal, RGB, or HSL color. "hsl(152, 80%, 80%)" is the default.

color_max
string representing the maximum value of the color range for the circles. The string can be either a hexadecimal, RGB, or HSL color. "hsl(228, 30%,40%)" is the default.

...
other arguments passing to circlepackeR.

Details

Plot diversity (q= 0,1,2,...) with self-defined hierarchy.

Value

A interactive plot showing the size of biodiversity in a hierarchy.

References

Chao, A., Chiu, C. H., & Jost, L. (2014). Unifying species diversity, phylogenetic diversity, functional diversity, and related similarity and differentiation measures through Hill numbers. Annual review of ecology, evolution, and systematics, 45, 297-324.

Qin, X. (2019). HierDpart: partitioning hierarchical diversity and differentiation across metrics and scales, from genes to ecosystems. R package version 0.5.0 https://cran.r-project.org/package=HierDpart.

Marcon, E., & Herault, B. (2015). entropart: An R package to measure and partition diversity. Journal of Statistical Software, 67(1), 1-26.

Jombart, T. (2008). adegenet: a R package for the multivariate analysis of genetic markers. Bioinformatics, 24(11), 1403-1405.

Froelich, Jerome. et al.2015. circlepackeR: R htmlwidget for circle packing visualization. R Package. https://github.com/jeromefroe/circlepackeR.

Author

qinxinghu@gmail.com

Examples

##example
data(Dprofile)

PlothieD(Dprofile[,-1],size = "Dq1")

#> Skipping install of 'circlepackeR' from a github remote, the SHA1 (f0a84d5e) has not changed since last install.
#>   Use `force = TRUE` to force installation

#> Skipping install of 'circlepackeR' from a github remote, the SHA1 (f0a84d5e) has not changed since last install.
#>   Use `force = TRUE` to force installation